Add support for class icons
Use the new `icon_path` field in the `GDExtensionClassCreationInfo4` struct, added to GDExtension in 4.4, to set the class icon if provided in the `[GodotClass]` attribute.

Since the icon must be set using the class creation info, we have to invoke the class configuration (so the `SetIcon` method is invoked) before calling `classdb_register_extension_class4`. This presents a problem because the class configuration will also register the properties, methods, and other members, but this can't be done before registering the extension class. In order to solve this, all binding registrations that need to happen after registering the class will be added to a queue to delay their execution until later.