[WIP] Dispose editor integration early
Some types require others when destroying their instances. For example, `DPITexture` uses `RenderingServer` in its destructor, but on the last deinitialization level the singletons have already been freed, so we need to make sure all C# instances have been disposed before that to prevent a crash. Luckily, in the editor integration we don't need to use any Godot types after the editor deinitialization level, it's still a bit of an ugly solution though.